requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
command-not-found==0.3
cryptography==2.3
cupshelpers==1.0
defer==1.0.6
defusedxml==0.5.0
distro==1.3.0
distro-info==0.18
Django==2.1.5
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
entrypoints==0.2.3.post3
gunicorn==19.7.1
httplib2==0.11.3
idna==2.7
keyring==15.1.0
keyrings.alt==3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
louis==3.6.0
macaroonbakery==1.1.4
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.1.0
olefile==0.46
pexpect==4.6.0
Pillow==5.2.0
protobuf==3.0.0
psycopg2==2.7.7
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.30.1
PyJWT==1.6.4
pymacaroons==0.13.0
PyNaCl==1.2.1
pyRFC3339==1.1
python-apt==1.7.0
python-debian==0.1.33
python3-openid==3.1.0
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.5.6
requests==2.20.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.15.0
six==1.11.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.23
usb-creator==0.3.3
virtualenv==16.4.3
wadllib==1.3.3
whitenoise==3.3.1
xkit==0.0.0